Agregados render de errores en productos, GET de productos sin Login

+++ src/controllers/carts.controller.js
@@ -19,13 +19,9 @@ const transporter = require('../utils/nodemailer.config')
 const createCart = async (req, res) => {
     const user = await userModel.findOne({_id: req.session.passport.user});
     const cart = await businessCarts.getByEmail(user.email)
-    if(cart){
-        logger.info("El carrito ya existe")
-    }
-    else{
+    if(!cart){
         const cart = {email: user.email, address: user.address, products: [], timestamp: Date.now()}
         await businessCarts.save(cart)
-        logger.info("Carrito creado")
     }
 }
 
@@ -53,13 +49,11 @@ const createCart = async (req, res) => {
 // }
 
 const deleteCart = async (req, res) => {
-    const { id_cart } = req.body
-    const cart = await businessCarts.getById(id_cart)
-    cart
-    ? (await businessCarts.deleteById(cart.id), 
-    res.redirect('/carrito'))
+    const { id_cart } = req.body;
+    await businessCarts.deleteById(id_cart);
+    res.redirect('/carrito');
     // res.status(200).json({ message: 'Carrito eliminado' }))
-    : res.status(400).json({ message: 'El carrito no existe' })
+    // : res.status(400).json({ message: 'El carrito no existe' })
 }
 
 // const deleteCart = async (req, res) => {
@@ -101,27 +95,30 @@ const updateCart = async (req, res) => {
     const user = await userModel.findOne({_id: req.session.passport.user});
     await createCart(req)
     let cart = await businessCarts.getByEmail(user.email);
-    // if(!cart){
-    //     cart = {email: user.email, address: user.address, products: [], timestamp: Date.now()}
-    //     await businessCarts.save(cart)
-    // }
-    let product = null
+    if(qty>0){
+    let product
     process.env.DATABASE === 'file' 
     ? product = await businessProds.getById(id_prod)
     : product = await businessProds.getNative(id_prod)
-    console.log(product)
-    if(product){
-        product._id = id_prod
-        product.id = crypto.randomBytes(10).toString('hex');
-        product.qty = parseInt(qty)
-        delete product.stock
-        cart.products.push(product)
-        await businessCarts.updateById(cart.id, {products: cart.products, timestamp: cart.timestamp})
-        // res.status(200).json({ message: 'Producto añadido al carrito' })
-        res.redirect('/carrito')
+        if(product){
+            product._id = id_prod
+            const prodCart = cart.products.find(p => p._id == id_prod)
+            cart.products = cart.products.filter(p => p !== prodCart)
+            product.id = crypto.randomBytes(10).toString('hex');
+            product.qty = parseInt(qty)
+            delete product.stock
+            cart.products.push(product)
+            await businessCarts.updateById(cart.id, {products: cart.products, timestamp: cart.timestamp})
+            // res.status(200).json({ message: 'Producto añadido al carrito' })
+            res.redirect('/carrito')
+        }
+        else{
+            res.render('pages/systemMessage', { message: 'El producto que intentas agregar al carrito no existe', success: false, href: 'productos' });
+            // res.status(400).json({ message: 'El producto no existe' })
+        }
     }
     else{
-        res.status(400).json({ message: 'El producto no existe' })
+        res.render('pages/systemMessage', { message: 'La cantidad a agregar del producto debe ser mayor a 0', success: false, href: 'productos' });
     }
 }
 
@@ -149,7 +146,8 @@ const deleteCartProduct = async (req, res) => {
     res.redirect('/carrito')
     }
     else{
-        res.status(400).json({ message: 'El producto seleccionado no existe en el carrito' })
+        res.render('pages/systemMessage', { message: 'El producto seleccionado no existe en el carrito', success: false, href: 'carrito' });
+        // res.status(400).json({ message: 'El producto seleccionado no existe en el carrito' })
     }
 }
 
@@ -172,43 +170,47 @@ const sendCart = async (req, res) => {
     const { id_cart, total } = req.body
     const user = await userModel.findOne({_id: req.session.passport.user});
     const cart = await businessCarts.getById(id_cart)
-    const order = { items: cart.products, total: total, date: formatDate(new Date()), state: 'Generada', buyer: cart.email }
-    let arrayItems = "";
-    let arrayItemsMsg = "";
-    let n;
-    for (n in cart.products) {
-    arrayItems += `<div style="color: #2bf8bb;">
-                        <strong><u><p style="color: #2bf8bb;">${cart.products[n].title}</p></u></strong>            
-                        <img src=${cart.products[n].thumbnail} width="50" height="50" style="color: #2bf8bb;" alt="Imagen Producto"/><br>
-                        <span style="color: #4eaa93;">Cantidad: ${cart.products[n].qty}</span>
-                        <h5 style="color: #4eaa93;">Precio Unitario: ${cart.products[n].price} U$S</h5>
-                        <h4 style="color: #2bf8bb;">Total Producto: ${cart.products[n].qty * cart.products[n].price} U$S</h4>
-                        <br>
-                    </div>`;
-    arrayItemsMsg += `
-     • ${cart.products[n].title}           
-     Cantidad: ${cart.products[n].qty}
-     Precio Unitario: ${cart.products[n].price} U$S
-     Total Producto: ${cart.products[n].qty * cart.products[n].price} U$S\n`;
+    if(cart){
+        const order = { items: cart.products, total: total, date: formatDate(new Date()), state: 'Generada', buyer: cart.email }
+        let arrayItems = "";
+        let arrayItemsMsg = "";
+        let n;
+        for (n in cart.products) {
+        arrayItems += `<div style="color: #2bf8bb;">
+                            <strong><u><p style="color: #2bf8bb;">${cart.products[n].title}</p></u></strong>            
+                            <img src=${cart.products[n].thumbnail} width="50" height="50" style="color: #2bf8bb;" alt="Imagen Producto"/><br>
+                            <span style="color: #4eaa93;">Cantidad: ${cart.products[n].qty}</span>
+                            <h5 style="color: #4eaa93;">Precio Unitario: ${cart.products[n].price} U$S</h5>
+                            <h4 style="color: #2bf8bb;">Total Producto: ${cart.products[n].qty * cart.products[n].price} U$S</h4>
+                            <br>
+                        </div>`;
+        arrayItemsMsg += `
+        • ${cart.products[n].title}           
+        Cantidad: ${cart.products[n].qty}
+        Precio Unitario: ${cart.products[n].price} U$S
+        Total Producto: ${cart.products[n].qty * cart.products[n].price} U$S\n`;
+        }
+        const mailOptions =  {
+            from: `${user.email}`,
+            to: mail,
+            subject: `Nueva orden de compra de: ${user.name}`,
+            html: `<div style="background-color:black;"><br>
+                    <h1 style="color: #2bf8bb;">&nbsp&nbsp&nbsp Pedido de ${user.name}:</h1>
+                    <ul>${arrayItems}</ul>
+                    <h2 style="color: #2bf8bb;">&nbsp&nbsp&nbspTotal: ${total} U$S</h2><br>
+                    </div><br>`
+        }
+        // await transporter.sendMail(mailOptions)
+        // await sendMsg(`Hola ${user.name}!, tu pedido N° ${cart.timestamp} fue recibido y se encuentra en proceso!`,'+14793365162',process.env.PHONE)
+        // await sendMsg(`Pedido de ${user.name}\n ${arrayItemsMsg}\n Total: ${total} U$S `,'whatsapp:+14155238886',`whatsapp:${process.env.WHATSAPP_PHONE}`)
+        await businessOrders.save(order)
+        await businessCarts.deleteById(id_cart)
+        res.render('pages/systemMessage', { message: `Se ha enviado tu orden de compra, podés consultar su estado en /ordenes`, success: true, href: 'productos' });
+        // res.status(200).json({ message: 'Pedido enviado' })
     }
-    const mailOptions =  {
-        from: `${user.email}`,
-        to: mail,
-        subject: `Nueva orden de compra de: ${user.name}`,
-        html: `<div style="background-color:black;"><br>
-                <h1 style="color: #2bf8bb;">&nbsp&nbsp&nbsp Pedido de ${user.name}:</h1>
-                <ul>${arrayItems}</ul>
-                <h2 style="color: #2bf8bb;">&nbsp&nbsp&nbspTotal: ${total} U$S</h2><br>
-                </div><br>`
+    else{
+        res.render('pages/systemMessage', { message: `Ha surgido un error al intentar enviar tu orden, ingresa a tu carrito e intentalo nuevamente`, success: false, href: 'carrito' });
     }
-    // await transporter.sendMail(mailOptions)
-    // await sendMsg(`Hola ${user.name}!, tu pedido N° ${cart.timestamp} fue recibido y se encuentra en proceso!`,'+14793365162',process.env.PHONE)
-    // await sendMsg(`Pedido de ${user.name}\n ${arrayItemsMsg}\n Total: ${total} U$S `,'whatsapp:+14155238886',`whatsapp:${process.env.WHATSAPP_PHONE}`)
-    // console.log(await ordersModel.create(order))
-    await businessOrders.save(order)
-    await businessCarts.deleteById(id_cart)
-    res.redirect('/productos')
-    // res.status(200).json({ message: 'Pedido enviado' })
 }
 
 module.exports = {
